% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{edhec}
\alias{edhec}
\title{edhec documentation here}
\format{An object of class \code{xts} (inherits from \code{zoo}) with 152 rows and 13 columns.}
\usage{
edhec
}
\description{
edhec documentation here
}
\keyword{datasets}
