% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Measures.R
\name{ES.hist}
\alias{ES.hist}
\title{Compute sample Expected Shortfall}
\usage{
ES.hist(data, ..., alpha.ES = 0.05)
}
\arguments{
\item{data}{Vector of data}

\item{alpha}{Tail Probability}
}
\value{
sample ES
}
\description{
Compute sample Expected Shortfall
}
\examples{
ES(rnorm(10))
}
\author{
Xin Chen, \email{chenx26@uw.edu}
}
