% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Measures.R
\name{ES}
\alias{ES}
\title{Compute sample Expected Shortfall}
\usage{
ES(data, alpha = 0.1)
}
\arguments{
\item{data}{Vector of data}

\item{alpha}{Tail Probability}
}
\value{
sample ES
}
\description{
Compute sample Expected Shortfall
}
\examples{
ES(rnorm(10))
}

