% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello.name}
\alias{hello.name}
\title{Print "Hello, name"}
\usage{
hello.name(name)
}
\arguments{
\item{name}{The name to say hello to}
}
\value{
None
}
\description{
Print "Hello, name"
}
\examples{
hello.name("world")
}

