% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello2}
\alias{hello2}
\title{Print "Greetings, yourname"}
\usage{
hello2(yourname)
}
\arguments{
\item{yourname}{The name to greet}
}
\value{
None
}
\description{
Print "Greetings, yourname"
}
\examples{
hello2("world")
}

