% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InfluenceFunctions.R
\name{LPM.IF}
\alias{LPM.IF}
\title{Compute the influence function of LPM}
\usage{
LPM.IF(data, ..., const = 0, k = 1)
}
\arguments{
\item{data}{vector of data}

\item{...}{other parameters}

\item{const}{the constant threshold}

\item{k}{the order of LPM, can only be 1 or 2}
}
\value{
inlfuence function of LPM
}
\description{
Compute the influence function of LPM
}
\examples{
LPM.IF(rnorm(10), const = 0.1, k=2)
}
\author{
Xin Chen, \email{chenx26@uw.edu}
}
