% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Measures.R
\name{LPM}
\alias{LPM}
\title{Compute sample lower partial moment}
\usage{
LPM(data, ..., const = 0, k = 1)
}
\arguments{
\item{data}{vector of data}

\item{...}{other parameters}

\item{const}{the constant threshold}

\item{k}{order of the LPM to be computed}
}
\value{
sample lower partial moment
}
\description{
Compute sample lower partial moment
}
\examples{
LPM(rnorm(10),const = -0.1)
}
\author{
Xin Chen, \email{chenx26@uw.edu}
}
