% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{managers}
\alias{managers}
\title{managers documentation here}
\format{An object of class \code{xts} (inherits from \code{zoo}) with 132 rows and 10 columns.}
\usage{
managers
}
\description{
managers documentation here
}
\keyword{datasets}
