% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InfluenceFunctions.R
\name{OmegaRatio.IF}
\alias{OmegaRatio.IF}
\title{Compute influence function of Omega Ratio}
\usage{
OmegaRatio.IF(data, ..., const = 0)
}
\arguments{
\item{data}{vector of data}

\item{...}{other parameters}

\item{const}{the constant threshold}
}
\value{
IF of Omega
}
\description{
Compute influence function of Omega Ratio
}
\examples{
OmegaRatio.IF(rnorm(10), const = 0.1)
}
\author{
Xin Chen, \email{chenx26@uw.edu}
}

