% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Measures.R
\name{OmegaRatio}
\alias{OmegaRatio}
\title{Compute sample Omega Ratio}
\usage{
OmegaRatio(data, ..., const = 0)
}
\arguments{
\item{data}{vector of data}

\item{...}{other parameters}

\item{const}{the constant threshold}
}
\value{
sample Omega Ratio
}
\description{
Compute sample Omega Ratio
}
\examples{
OmegaRatio(rnorm(10),const = -0.1)
}
\author{
Xin Chen, \email{chenx26@uw.edu}
}
