% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{portfolio_bacon}
\alias{portfolio_bacon}
\title{portfolio_bacon documentation here}
\format{An object of class \code{xts} (inherits from \code{zoo}) with 24 rows and 2 columns.}
\usage{
portfolio_bacon
}
\description{
portfolio_bacon documentation here
}
\keyword{datasets}
