% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prices}
\alias{prices}
\title{prices documentation here}
\format{An object of class \code{zoo} with 2011 rows and 1 columns.}
\usage{
prices
}
\description{
prices documentation here
}
\keyword{datasets}
