% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printSE.R
\name{printSE}
\alias{printSE}
\title{Formatted output of the results from xxx.SE() functions}
\usage{
printSE(res, round.digit = 3, valonly = TRUE)
}
\arguments{
\item{res}{the results from the xx.SE() functions}

\item{round.digit}{the number of digits to round to}

\item{valonly}{if true, return only the data frame, otherwise print and return data frame}
}
\value{
data frame
}
\description{
Formatted output of the results from xxx.SE() functions
}
\examples{
data(edhec)
res = ES.SE(edhec, p=.95, method="historical",se.method = "IFiid")
printSE(res, round.digit = 4)
}
\author{
Xin Chen, \email{chenx26@uw.edu}
}
