% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LongRunVarianceEstimators.R
\name{sampleVariance.cor}
\alias{sampleVariance.cor}
\title{Naive Sample Variance of sample mean for correlated data}
\usage{
sampleVariance.cor(data)
}
\arguments{
\item{data}{The vector of the data}
}
\value{
Sample Variance of the sample mean for correlated data
}
\description{
Naive Sample Variance of sample mean for correlated data
}
\examples{
sampleVariance.cor(data)
}

