% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SE_AR.R
\name{SE.AR}
\alias{SE.AR}
\title{Compute long run variance by fitting AR model with AIC or BIC model selection}
\usage{
SE.AR(x, max.order = 10, criteria = c("AIC", "BIC"), all.results = FALSE)
}
\arguments{
\item{x}{vector of input data time series}

\item{max.order}{maximum order for AR model, default 10}

\item{criteria}{model selection criteria, "AIC" or "BIC"}

\item{all.results}{if TRUE, return all fitted model along with longrun variance, otherwise return lrvar only}
}
\value{
lrvar or a list with first element lrvar and second element a list of all fitted model
}
\description{
Compute long run variance by fitting AR model with AIC or BIC model selection
}
\examples{
SE.AR(rnorm(100))
}
