% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SE_IID.R
\name{SE.ES.iid}
\alias{SE.ES.iid}
\title{Function to compute the asymptotic standard error of the sample expected shortfall for the given IID data}
\usage{
SE.ES.iid(data, alpha = 0.05)
}
\arguments{
\item{data}{The vector of data}

\item{alpha}{The tail probability}
}
\value{
The asymptotic standard error of the sample expected shortfall
}
\description{
This function only works if the data is IID
}
\examples{
x=rnorm(10)
SE.ES.iid(x,alpha=0.01)
}

