% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLM_EN.R
\name{SE.glmnet_exp}
\alias{SE.glmnet_exp}
\title{Compute the standard error of the mean of the data using glmnet_exp}
\usage{
SE.glmnet_exp(data, ..., d = 7, alpha = 0.5, keep = 1,
  standardize = FALSE)
}
\arguments{
\item{data}{vector of data}

\item{d}{degree of polynomial}

\item{alpha}{weight for Elastic Net regularizer}

\item{keep}{what portion of sample spectral density to use for model fitting}
}
\value{
variance of the mean of the data
}
\description{
Compute the standard error of the mean of the data using glmnet_exp
}
