% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HW1981.R
\name{SE.HW1981}
\alias{SE.HW1981}
\title{Exact implementation of H&W 1981, compute the variance of the sample mean of the data}
\usage{
SE.HW1981(data, K = 25, d = 2)
}
\arguments{
\item{data}{Vector of data}

\item{K}{Number of periodograms to keep}

\item{d}{Order of the polynomial fit}
}
\value{
Variance of the sample mean. To get asymptotic variance, multiply by length(data)
}
\description{
Exact implementation of H&W 1981, compute the variance of the sample mean of the data
}
\examples{
SE.HW1981(rnorm(10))
}
