% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SE_Mean.R
\name{SE.Mean.boot.iid}
\alias{SE.Mean.boot.iid}
\title{Compute standard error of mean via bootstrapping for iid data}
\usage{
SE.Mean.boot.iid(data, ...)
}
\arguments{
\item{data}{vector of data}

\item{...}{parameters passed from upper calls}
}
\value{
SE of mean
}
\description{
Compute standard error of mean via bootstrapping for iid data
}
\examples{
SE.Mean.boot.iid(rnorm(10))
}

