% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SE_Mean.R
\name{SE.Mean.iid}
\alias{SE.Mean.iid}
\title{Standard Error of Mean for iid data}
\usage{
SE.Mean.iid(data)
}
\arguments{
\item{data}{vector of data}
}
\value{
SE of Mean
}
\description{
Standard Error of Mean for iid data
}
\examples{
SE.Mean.iid(rnorm(10))
}

