% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SE_IID.R
\name{SE.mu.iid}
\alias{SE.mu.iid}
\title{Function to compute the asymptotic standard error of the sample mean for the given IID data}
\usage{
SE.mu.iid(data)
}
\arguments{
\item{data}{The vector of data}
}
\value{
The asymptotic standard error of the sample mean
}
\description{
This function only works if the data is IID
}
\examples{
x=rnorm(10)
SE.mu.iid(x)
}

