% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SE_pspectrum.R
\name{SE.pspectrum}
\alias{SE.pspectrum}
\title{Use pspectrum function to compute long run variance}
\usage{
SE.pspectrum(x, ..., verbose = FALSE, plot = FALSE)
}
\arguments{
\item{x}{data}

\item{...}{other parameters}

\item{verbose}{verbose or not}

\item{plot}{plot or not}
}
\value{
long run variance of x
}
\description{
Use pspectrum function to compute long run variance
}
