% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SE_SD.R
\name{SE.SD.boot.cor}
\alias{SE.SD.boot.cor}
\title{Compute standard error of SD via bootstrapping for data (possibily serially correlated)}
\usage{
SE.SD.boot.cor(data, ..., segment.length = length(data)/5)
}
\arguments{
\item{data}{vector of data}

\item{...}{parameters passed from upper calls}

\item{segment.length}{the length of the fixed block for bootstrapping}
}
\value{
SE of SD
}
\description{
Compute standard error of SD via bootstrapping for data (possibily serially correlated)
}
\examples{
SE.SD.boot.cor(rnorm(100))
}

