% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SE_SD.R
\name{SE.SD.boot.iid}
\alias{SE.SD.boot.iid}
\title{Compute standard error of SD via bootstrapping for iid data}
\usage{
SE.SD.boot.iid(data, ...)
}
\arguments{
\item{data}{vector of data}

\item{...}{parameters passed from upper calls}
}
\value{
SE of SD
}
\description{
Compute standard error of SD via bootstrapping for iid data
}
\examples{
SE.SD.boot.iid(rnorm(10))
}

