% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SE_SD.R
\name{SE.SD.boot.iid.xts}
\alias{SE.SD.boot.iid.xts}
\title{Wrapper function to compute the standard error(s) of the SD for an xts object
using iid bootstrapping}
\usage{
SE.SD.boot.iid.xts(x, ...)
}
\arguments{
\item{x}{the xts object}

\item{...}{other parameters}
}
\value{
standard error(s) of the xts object
}
\description{
The operation is performed column wise.
}
\examples{
data(edhec)
SE.SD.boot.iid.xts(edhec)
}

