% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SE_SD.R
\name{SE.SD.cor}
\alias{SE.SD.cor}
\title{Compute the standard error of SD estimator for correlated data vector using
GLM-EN method}
\usage{
SE.SD.cor(x, d = 5, alpha.lasso = 0.5, keep = 1)
}
\arguments{
\item{x}{the vector for the data}

\item{d}{the maximum order of the polynomial}

\item{alpha.lasso}{the weight for lasso vs elastic net}

\item{keep}{the portion of peridograms to use to fit the polynomial}
}
\value{
the SE of the SD estimtor for the data
}
\description{
Compute the standard error of SD estimator for correlated data vector using
GLM-EN method
}

