% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SE_SD.R
\name{SE.SD.cor.xts}
\alias{SE.SD.cor.xts}
\title{Wrapper function to compute the standard error(s) of the SD for an xts object
using GLM-EN method}
\usage{
SE.SD.cor.xts(x)
}
\arguments{
\item{x}{the xts object}
}
\value{
standard error(s) of the xts object
}
\description{
The operation is performed column wise.
}

