% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SE_SD.R
\name{SE.SD.iid.xts}
\alias{SE.SD.iid.xts}
\title{Wrapper function to compute the standard error(s) of the standard deviation for an xts object}
\usage{
SE.SD.iid.xts(x)
}
\arguments{
\item{x}{the xts object}
}
\value{
standard error(s) of the xts object
}
\description{
The operation is performed column wise.
}
\examples{
data(edhec)
SE.SD.iid.xts(edhec)
}

