% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SE_SD.R
\name{SE.SD}
\alias{SE.SD}
\title{Wrapper function to compute the standard error for standard deviation estimator}
\usage{
SE.SD(data, ..., se.method = c("none", "IFiid", "IFcor", "BOOTiid",
  "BOOTcor"))
}
\arguments{
\item{data}{xts object}

\item{...}{extra arugements to be passed to lower level functions, see details}

\item{se.method}{a character string indicating which method should be used to compute
the standard error of the estimated standard deviation. One of \code{"none"} (default),
\code{"IFiid"}, \code{"IFcor"}, \code{"BOOTiid"}, \code{"BOOTcor"}. Currely, only \code{"IFiid"}
is implemented.}
}
\value{
SE of SD
}
\description{
Wrapper function to compute the standard error for standard deviation estimator
}
\examples{
data(edhec)
SE.SD(edhec,se.method="IFiid")
}

