% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SE_SoR.R
\name{SE.SoR.boot.iid}
\alias{SE.SoR.boot.iid}
\title{Compute standard error of SoR via bootstrapping for iid data}
\usage{
SE.SoR.boot.iid(data, ..., rf = 0)
}
\arguments{
\item{data}{vector of data}

\item{...}{parameters passed from upper calls}
}
\value{
SE of SoR
}
\description{
Compute standard error of SoR via bootstrapping for iid data
}
\examples{
SE.SoR.boot.iid(rnorm(10))
}

