% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SE_SoR.R
\name{SE.SoR.iid}
\alias{SE.SoR.iid}
\title{Standard Error of SoR for iid data}
\usage{
SE.SoR.iid(data, rf = 0)
}
\arguments{
\item{data}{vector of data}

\item{rf}{risk free interest rate}
}
\value{
SE of SoR
}
\description{
Standard Error of SoR for iid data
}
\examples{
SE.SoR.iid(rnorm(10))
}

