% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SE_SoR.R
\name{SE.SoR.iid.xts}
\alias{SE.SoR.iid.xts}
\title{Wrapper function to compute the standard error(s) of the Sharpe Raio for an xts object}
\usage{
SE.SoR.iid.xts(x, ...)
}
\arguments{
\item{x}{the xts object}
}
\value{
standard error(s) of the xts object
}
\description{
The operation is performed column wise.
}
\examples{
data(edhec)
SE.SoR.iid.xts(edhec)
}

