% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SE_SoR.R
\name{SE.SoR}
\alias{SE.SoR}
\title{Wrapper function to compute the standard error for Sortino Ratio estimator}
\usage{
SE.SoR(data, ..., se.method = c("none", "IFiid", "IFcor", "BOOTiid",
  "BOOTcor"))
}
\arguments{
\item{data}{xts object}

\item{...}{extra arugements to be passed to lower level functions, see details}

\item{se.method}{a character string indicating which method should be used to compute
the standard error of the estimated Sharpe Ratio. One of \code{"none"} (default),
\code{"IFiid"}, \code{"IFcor"}, \code{"BOOTiid"}, \code{"BOOTcor"}. Currely, only \code{"IFiid"}
is implemented.}
}
\value{
SE of SoR
}
\description{
Wrapper function to compute the standard error for Sortino Ratio estimator
}
\examples{
data(edhec)
SE.SoR(edhec,se.method="IFiid")
}

