% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SE_SR.R
\name{SE.SR.boot.iid}
\alias{SE.SR.boot.iid}
\title{Compute standard error of SR via bootstrapping for iid data}
\usage{
SE.SR.boot.iid(data, ...)
}
\arguments{
\item{data}{vector of data}

\item{...}{parameters passed from upper calls}
}
\value{
SE of SR
}
\description{
Compute standard error of SR via bootstrapping for iid data
}
\examples{
SE.SR.boot.iid(rnorm(10))
}

