% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SE_IID.R, R/SE_SR.R
\name{SE.SR.iid}
\alias{SE.SR.iid}
\title{Function to compute the asymptotic standard error of the sample sharpe ratio for the given IID data}
\usage{
SE.SR.iid(data)

SE.SR.iid(data)
}
\arguments{
\item{data}{The vector of data}

\item{data}{vector of data}
}
\value{
The asymptotic standard error of the sample sharpe ratio

SE of SR
}
\description{
This function only works if the data is IID

Standard Error of SR for iid data
}
\examples{
x=rnorm(10)
SE.SR.iid(x)
SE.SR.iid(rnorm(10))
}

