% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SE_SR.R
\name{SE.SR.iid.xts}
\alias{SE.SR.iid.xts}
\title{Wrapper function to compute the standard error(s) of the Sharpe Raio for an xts object}
\usage{
SE.SR.iid.xts(x, ..., rf = 0)
}
\arguments{
\item{x}{the xts object}

\item{...}{other parameters}

\item{rf}{risk free interest rate}
}
\value{
standard error(s) of the xts object
}
\description{
The operation is performed column wise.
}
\examples{
data(edhec)
SE.SR.iid.xts(edhec)
}

