% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SE_STARR.R
\name{SE.STARR.iid}
\alias{SE.STARR.iid}
\title{Standard Error of STARR for iid data}
\usage{
SE.STARR.iid(data, alpha = 0.05, rf = 0)
}
\arguments{
\item{data}{vector of data}

\item{alpha}{tail probability}

\item{rf}{risk free interest rate}
}
\value{
SE of STARR
}
\description{
Standard Error of STARR for iid data
}
\examples{
SE.STARR.iid(rnorm(10))
}

