% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SE_STARR.R
\name{SE.STARR.iid.xts}
\alias{SE.STARR.iid.xts}
\title{Wrapper function to compute the standard error(s) of the STARR for an xts object}
\usage{
SE.STARR.iid.xts(x, alpha = 0.05, rf = 0)
}
\arguments{
\item{x}{the xts object}

\item{alpha}{tail probability}

\item{rf}{risk free interest rate}
}
\value{
standard error(s) of the xts object
}
\description{
The operation is performed column wise.
}
\examples{
data(edhec)
SE.STARR.iid.xts(edhec)
}

