% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SE_VaR.R
\name{SE.VaR.boot.iid}
\alias{SE.VaR.boot.iid}
\title{Compute standard error of VaR via bootstrapping for iid data}
\usage{
SE.VaR.boot.iid(data, ..., alpha = 0.05)
}
\arguments{
\item{data}{vector of data}

\item{...}{parameters passed from upper calls}
}
\value{
SE of VaR
}
\description{
Compute standard error of VaR via bootstrapping for iid data
}
\examples{
SE.VaR.boot.iid(rnorm(10))
}

