% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Measures.R
\name{SoR}
\alias{SoR}
\title{Compute sample Sortino Ratio}
\usage{
SoR(data, rf = 0)
}
\arguments{
\item{data}{vector of data}

\item{rf}{risk-free rate}
}
\value{
Sample SoR
}
\description{
Compute sample Sortino Ratio
}
\examples{
SoR(rnorm(10))
}

