% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampleMomentConditionFunctions.R
\name{SR.SMCF}
\alias{SR.SMCF}
\title{Compute the sample moment condition functions of SR from data}
\usage{
SR.SMCF(data)
}
\arguments{
\item{data}{Vector of data}
}
\value{
2 by N matrix of Sample Moment Condition Functions
}
\description{
Compute the sample moment condition functions of SR from data
}
\examples{
SR.SMCF(rnorm(10))
}

