% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InfluenceFunctions.R
\name{SSD.IF}
\alias{SSD.IF}
\title{Influence Function of Semi-Standard Deviation using mean as the threshold}
\usage{
\method{SSD}{IF}(data, ..., rf = 0)
}
\arguments{
\item{data}{Vector of the data}

\item{...}{other parameters}
}
\value{
IF of SSD
}
\description{
Influence Function of Semi-Standard Deviation using mean as the threshold
}
\examples{
SSD.IF(rnorm(10))
}
\author{
Xin Chen, \email{chenx26@uw.edu}
}
