% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{weights}
\alias{weights}
\title{weights documentation here}
\format{An object of class \code{xts} (inherits from \code{zoo}) with 8 rows and 11 columns.}
\usage{
weights
}
\description{
weights documentation here
}
\keyword{datasets}
