% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SE_ES.R, R/SE_IID.R
\name{SE.ES.iid}
\alias{SE.ES.iid}
\title{Standard Error of ES for iid data}
\usage{
SE.ES.iid(data, alpha = 0.05)

SE.ES.iid(data, alpha = 0.05)
}
\arguments{
\item{data}{vector of data}

\item{alpha}{The tail probability}

\item{data}{The vector of data}
}
\value{
SE of ES

The asymptotic standard error of the sample expected shortfall
}
\description{
Standard Error of ES for iid data

This function only works if the data is IID
}
\examples{
SE.ES.iid(rnorm(10))
x=rnorm(10)
SE.ES.iid(x,alpha=0.01)
}

