% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SE_SD.R
\name{SE.SD.iid}
\alias{SE.SD.iid}
\title{Standard Error of SD for iid data}
\usage{
SE.SD.iid(data)
}
\arguments{
\item{data}{vector of inputs}
}
\value{
SE of SD
}
\description{
Standard Error of SD for iid data
}
\examples{
SE.SD.iid(rnorm(10))
}

