% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SE_IID.R, R/SE_SD.R
\name{SE.SD.iid}
\alias{SE.SD.iid}
\title{Function to compute the asymptotic standard error of the sample standard deviation for the given IID data}
\usage{
SE.SD.iid(data)

SE.SD.iid(data)
}
\arguments{
\item{data}{The vector of data}

\item{data}{vector of inputs}
}
\value{
The asymptotic standard error of the sample standard deviation

SE of SD
}
\description{
This function only works if the data is IID

Standard Error of SD for iid data
}
\examples{
x=rnorm(10)
SE.SD.iid(x)
SE.SD.iid(rnorm(10))
}

