% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SE_IID.R, R/SE_SD.R
\name{SE.SD.iid.xts}
\alias{SE.SD.iid.xts}
\title{Asymptotic Standard Error of standard deviation for xts object}
\usage{
SE.SD.iid.xts(x)

SE.SD.iid.xts(x)
}
\arguments{
\item{x}{xts object}

\item{na.rm}{whether NAs should be omitted or not}

\item{x}{the xts object}
}
\value{
Asymptotic SE of SD for the xts object

standard error(s) of the xts object
}
\description{
Asymptotic Standard Error of standard deviation for xts object

The operation is performed column wise.
}
\examples{
SE.SD.iid.xts(rnorm(10))
data(edhec)
SE.SD.iid.xts(edhec)
}

