% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SE_IID.R
\name{SE.SoR.iid}
\alias{SE.SoR.iid}
\title{Function to compute the asymptotic standard error of the sample Sortino Ratio for the given IID data}
\usage{
SE.SoR.iid(data, rf = 0)
}
\arguments{
\item{data}{The vector of data}

\item{rf}{The risk free rate}
}
\value{
The asymptotic standard error of the sample Sortino Ratio
}
\description{
This function only works if the data is IID
}
\examples{
x=rnorm(10)
SE.SoR.iid(x)
}

