% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SE_IID.R, R/SE_SoR.R
\name{SE.SoR.iid}
\alias{SE.SoR.iid}
\title{Function to compute the asymptotic standard error of the sample Sortino Ratio for the given IID data}
\usage{
SE.SoR.iid(data, rf = 0)

SE.SoR.iid(data, rf = 0)
}
\arguments{
\item{data}{The vector of data}

\item{rf}{The risk free rate}

\item{data}{vector of data}
}
\value{
The asymptotic standard error of the sample Sortino Ratio

SE of SoR
}
\description{
This function only works if the data is IID

Standard Error of SoR for iid data
}
\examples{
x=rnorm(10)
SE.SoR.iid(x)
SE.SoR.iid(rnorm(10))
}

