% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SE_SR.R
\name{SE.SR.iid}
\alias{SE.SR.iid}
\title{Standard Error of SR for iid data}
\usage{
SE.SR.iid(data, rf = 0)
}
\arguments{
\item{data}{vector of data}

\item{rf}{risk free interest rate}
}
\value{
SE of SR
}
\description{
Standard Error of SR for iid data
}
\examples{
SE.SR.iid(rnorm(10))
}

