% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SE_IID.R
\name{SE.SR.iid}
\alias{SE.SR.iid}
\title{Function to compute the asymptotic standard error of the sample sharpe ratio for the given IID data}
\usage{
SE.SR.iid(data, rf = 0)
}
\arguments{
\item{data}{The vector of data}
}
\value{
The asymptotic standard error of the sample sharpe ratio
}
\description{
This function only works if the data is IID
}
\examples{
x=rnorm(10)
SE.SR.iid(x)
}

