% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SE_STARR.R
\name{SE.STARR.boot.iid}
\alias{SE.STARR.boot.iid}
\title{Compute standard error of STARR via bootstrapping for iid data}
\usage{
SE.STARR.boot.iid(data, ..., alpha = 0.05, rf = 0)
}
\arguments{
\item{data}{vector of data}

\item{...}{parameters passed from upper calls}

\item{alpha}{tail probability}

\item{rf}{risk free interest rate}
}
\value{
SE of STARR
}
\description{
Compute standard error of STARR via bootstrapping for iid data
}
\examples{
SE.STARR.boot.iid(rnorm(10))
}

