% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SE_IID.R, R/SE_STARR.R
\name{SE.STARR.iid}
\alias{SE.STARR.iid}
\title{Function to compute the asymptotic standard error of the sample STARR Ratio for the given IID data}
\usage{
SE.STARR.iid(data, alpha = 0.05, rf = 0)

SE.STARR.iid(data, alpha = 0.05, rf = 0)
}
\arguments{
\item{data}{The vector of data}

\item{alpha}{The tail probability used to compute expected shortfall}

\item{rf}{The risk free rate}

\item{data}{vector of data}
}
\value{
The asymptotic standard error of the sample STARR Ratio

SE of STARR
}
\description{
This function only works if the data is IID

Standard Error of STARR for iid data
}
\examples{
x=rnorm(10)
SE.STARR.iid(x)
SE.STARR.iid(rnorm(10))
}

