% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SE_IID.R, R/SE_VaR.R
\name{SE.VaR.iid}
\alias{SE.VaR.iid}
\title{Function to compute the asymptotic standard error of the sample value-at-risk for the given IID data}
\usage{
SE.VaR.iid(data, alpha = 0.05)

SE.VaR.iid(data, alpha = 0.05)
}
\arguments{
\item{data}{The vector of data}

\item{alpha}{tail probability}

\item{data}{vector of data}
}
\value{
The asymptotic standard error of the sample value-at-risk

SE of VaR
}
\description{
This function only works if the data is IID

Standard Error of VaR for iid data
}
\examples{
x=rnorm(10)
SE.VaR.iid(x)
SE.VaR.iid(rnorm(10))
}

