% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SE_VaR.R
\name{SE.VaR.iid}
\alias{SE.VaR.iid}
\title{Standard Error of VaR for iid data}
\usage{
SE.VaR.iid(data, alpha = 0.05)
}
\arguments{
\item{data}{vector of data}

\item{alpha}{tail probability}
}
\value{
SE of VaR
}
\description{
Standard Error of VaR for iid data
}
\examples{
SE.VaR.iid(rnorm(10))
}

